Fix for infinite loop when processing a chunked response
When StreamContent processes a chunked response, the
System.Net._InputNetworkStreamWrapper reads all chunks.
When processing the last chunk its length is 0.
SteamContent receives back the number of read bytes (0)
and continues to read the stream which returns 0 bytes
read etc...
Return -1 from the stream instead to indicate the end of
chunked processing.

Fix #984

Signed-off-by: Klaus Vancamelbeke <klaus@proles.be>