Fix for StreamContent not reading the entire NetworkStream
StreamContent uses the Content-Length header to decide when all is read.
This commit ensures that the headers are present in StreamContent, utilizes
the content length if present. If content length is not provided (i.e. chunked
response) the InputNetworkStreamWrapper uses a new IKnowWhenDone
interface to let StreamContent know when the last chunk (length == 0) is
processed.

Signed-off-by: Klaus Vancamelbeke (klaus@proles.be)