HttpContent.ReadAsString should only read applicable bytes instead of entire buffer
When using ReadAsString the current buffer is encoded to a string. It does not use MemoryStream.ToArray to prevent a copy in memory. The excess bytes in the buffer
are treated as well. This commit provides a way to set the actual read bytes on the
HttpContent and thus allows to set the real length of the buffer.

Signed-off-by: Klaus Vancamelbeke (klaus@proles.be)